+++ src/App.vue
@@ -1,12 +1,15 @@
 <template>
   <div id="q-app">
     <router-view />
+    <offline logo="../statics/trackit.png"/>
   </div>
 </template>
 
 <script>
+import Offline from './components/Offline'
 export default {
-  name: 'App'
+  name: 'App',
+  components: { Offline }
 }
 </script>