+++ src/components/Device.vue
@@ -59,7 +59,7 @@ export default {
       if (!this.activeDevicesID.includes(this.device.id)) {
         this.setActiveDevice()
       }
-      this.isDeviceWatched ? this.$emit('update:watch-by-id', null) : this.$emit('update:watch-by-id', this.device.id)
+      setTimeout(() => { this.isDeviceWatched ? this.$emit('update:watch-by-id', null) : this.$emit('update:watch-by-id', this.device.id) }, 500)
       setTimeout(() => { this.watchTooltip = false }, 500)
     }
   }