+++ src/components/Queue.vue
@@ -2,53 +2,33 @@
   <div id="queue" class="absolute-bottom-left absolute-bottom-right">
     <q-tabs v-model="selected" no-pane-border position="bottom" color="dark">
       <template v-for="(deviceID) in activeDevicesID">
-        <q-tab-pane :style="[{height:'20vh'}]" class="no-padding" :key="`tab-pane-${deviceID}`" :name="deviceID.toString()">
+        <q-tab-pane :style="[{height:'20vh', position: 'relative'}]" class="no-padding" :key="`tab-pane-${deviceID}`" :name="deviceID.toString()">
           <div class="no-messages text-center" v-if="!messages[deviceID].length">
             <div class="text-white" style="font-size: 3rem;">
               <div>No messages</div>
               <div style="font-size: 1.5rem;">or position is empty</div>
             </div>
-            <q-btn color="dark" v-if="isAdmin" @click="sendInitMessages(deviceID)">Send message</q-btn>
+            <q-btn color="dark" v-if="isAdmin && mode === 1" @click="sendInitMessages(deviceID)">Send message</q-btn>
           </div>
-          <div :style="{height: '100%'}" class="table__wrapper" v-if="messages[deviceID].length">
-            <table class="q-table highlight striped-odd cell-separator compact full-width text-white">
-              <thead>
-              <tr>
-                <th>Lat</th>
-                <th>Lon</th>
-                <th>Alt</th>
-                <th>Direction</th>
-                <th>Satellites</th>
-                <th>Speed</th>
-                <th>Time</th>
-                <th>Params</th>
-              </tr>
-              </thead>
-              <tbody>
-              <tr v-for="(message, index) in messages[deviceID]" :key="index" @click="selectedMessageIndex = index, $refs.messageViewer.show()" class="cursor-pointer">
-                <td data-th="Lat">{{message['position.latitude']}}</td>
-                <td data-th="Lon">{{message['position.longitude']}}</td>
-                <td data-th="Alt">{{message['position.altitude']}}</td>
-                <td data-th="Direction">{{message['position.direction']}}</td>
-                <td data-th="Satellites">{{message['position.satellites']}}</td>
-                <td data-th="Speed">{{message['position.speed']}}</td>
-                <td data-th="Time">{{getTime(message['timestamp'] * 1000)}}</td>
-                <td data-th="Params">{{getEtc(message)}}</td>
-              </tr>
-              </tbody>
-            </table>
+          <div class="table__wrapper" v-if="messages[deviceID].length">
+            <messages
+              :messages="messages[deviceID]"
+              :mode="mode"
+              :item="devices[deviceID]"
+              :activeId="deviceID"
+              :limit="2000"
+              :date="date"
+            ></messages>
           </div>
         </q-tab-pane>
         <q-tab :key="`tab-${deviceID}`" :label="getNameById(deviceID)" slot="title" :name="deviceID.toString()" />
       </template>
     </q-tabs>
-    <message-viewer ref="messageViewer" :message="typeof selectedMessageIndex !== 'undefined' ? messages[selected][selectedMessageIndex] : {}" inverted @close="selectedMessageIndex = undefined"></message-viewer>
   </div>
 </template>
 
 <script>
-import moment from 'moment'
-import MessageViewer from './MessageViewer'
+import Messages from './Messages'
 export default {
   name: 'Queue',
   props: [
@@ -56,11 +36,12 @@ export default {
     'activeDevicesID',
     'devices',
     'isAdmin',
-    'telemetryDeviceId'
+    'telemetryDeviceId',
+    'mode',
+    'date'
   ],
   data () {
     return {
-      selectedMessageIndex: undefined,
       selected: this.activeDevicesID.length ? this.activeDevicesID[0].toString() : '',
       telemetryOfActiveDevices: this.devices.filter(device => this.activeDevicesID.includes(device.id)).reduce((res, device) => {
         res[device.id] = device.telemetry ? device.telemetry : null
@@ -69,26 +50,8 @@ export default {
     }
   },
   methods: {
-    getEtc (msg) {
-      let buf = Object.assign({}, msg)
-      delete buf['position.latitude']
-      delete buf['position.longitude']
-      delete buf['position.altitude']
-      delete buf['position.direction']
-      delete buf['position.satellites']
-      delete buf['position.speed']
-      delete buf['timestamp']
-      delete buf['device_id']
-      if (!Object.keys(buf).length) {
-        return '*Empty*'
-      }
-      return JSON.stringify(buf)
-    },
     getNameById (id) {
       return this.devices.filter(device => device.id === id)[0].name || `&lt;#${id}&gt;`
-    },
-    getTime (ts) {
-      return moment(ts).format('L HH:mm:ss')
     },
     sendInitMessages (id) {
       this.$emit('send', id)
@@ -122,7 +85,7 @@ export default {
       }
     }
   },
-  components: {MessageViewer}
+  components: {Messages}
 }
 </script>
 
@@ -137,21 +100,7 @@ export default {
       overflow auto
       font-size 80%
       transition all .5s ease-in-out
-      thead
-        background-color: rgba(255, 255, 255, .3);
-        tr
-          height inherit
-          th
-            padding .1rem .3rem
-            text-align left
-            color white
-      tbody
-        tr
-          &:nth-child(odd)
-            background-color rgba(0,0,0,0.1)
-          td
-            height inherit
-            padding .1rem .3rem
+      height 100%
     .no-messages
       min-height: 140px;
       opacity: 0.7;