+++ src/pages/Login.vue
@@ -6,7 +6,7 @@
         Track it!
       </div>
     </div>
-    <div v-if="!$route.params.token && !offline">
+    <div v-if="!$route.params.token">
       <div class="login-card shadow-4 bg-white column items-center justify-center no-wrap">
         <p class="text-center">Track your devices on the map.</p>
         <div class="row full-width">
@@ -38,15 +38,6 @@
         </div>
       </div>
     </div>
-    <div v-else-if="offline" class="login-card shadow-4 bg-white">
-      <div class="column items-center justify-center no-wrap text-grey-6 uppercase" style="font-size: 10vmax;">
-        Offline
-      </div>
-      <q-progress indeterminate color="grey-6" style="width: 100%; height: 3px" />
-      <div class="text-center text-grey-6 uppercase">
-        waiting for reconnection
-      </div>
-    </div>
     <div v-else>
       <div class="login-card shadow-4 bg-white column items-center justify-center no-wrap">
         <q-progress indeterminate color="positive" style="width: 100%; height: 45px" />
@@ -56,8 +47,6 @@
 </template>
 
 <script>
-import { mapState, mapActions } from 'vuex'
-
 export default {
   data () {
     return {
@@ -66,7 +55,6 @@ export default {
     }
   },
   computed: {
-    ...mapState(['offline']),
     model: {
       get () {
         return this.token
@@ -77,7 +65,6 @@ export default {
     }
   },
   methods: {
-    ...mapActions(['checkConnection']),
     logIn () {
       this.$store.commit('setToken', this.token)
       this.$nextTick(() => { this.$router.push('/') })
@@ -115,26 +102,10 @@ export default {
       if (val.params && val.params.token) {
         this.autoLogin()
       }
-    },
-    offline (val) {
-      if (val) {
-        if (!this.offlineIntervalId) {
-          this.offlineIntervalId = setInterval(this.checkConnection, 5000)
-        }
-      } else {
-        clearInterval(this.offlineIntervalId)
-        this.checkHasToken()
-      }
     }
   },
   created () {
-    if (!this.offline) {
-      this.checkHasToken()
-    } else {
-      if (!this.offlineIntervalId) {
-        this.offlineIntervalId = setInterval(this.checkConnection, 5000)
-      }
-    }
+    this.checkHasToken()
   }
 }
 </script>